feat: add LevelKeyValueStore to inject AbstractLevelDOWN #49
