+++ package.json
@@ -33,7 +33,6 @@
     "memdown": "^5.1.0",
     "node-fetch": "^2.3.0",
     "ovm-compiler": "github:cryptoeconomicslab/ovm-compiler#v0.1.15",
-    "react-native-sqlite-storage": "^4.1.0",
     "socket.io": "^2.3.0",
     "socket.io-client": "^2.3.0",
     "tslib": "^1.10.0"
@@ -44,12 +43,11 @@
     "@types/levelup": "^3.1.1",
     "@types/libsodium-wrappers-sumo": "^0.7.3",
     "@types/memdown": "^3.0.0",
-    "@typescript-eslint/eslint-plugin": "^2.5.0",
-    "@typescript-eslint/parser": "^2.5.0",
     "@types/node-fetch": "^2.5.2",
-    "@types/react-native-sqlite-storage": "^3.3.1",
     "@types/socket.io": "^2.1.4",
     "@types/socket.io-client": "^1.4.32",
+    "@typescript-eslint/eslint-plugin": "^2.5.0",
+    "@typescript-eslint/parser": "^2.5.0",
     "eslint": "^6.6.0",
     "eslint-config-prettier": "^6.5.0",
     "eslint-plugin-prettier": "^3.1.1",
