+++ src/db/index.ts
@@ -7,5 +7,6 @@ export {
   createRangeWitnessStore
 } from './WitnessStore'
 export { InMemoryKeyValueStore } from './InMemoryKeyValueStore'
+export { LevelKeyValueStore } from './LevelKeyValueStore'
 export { IndexedDbKeyValueStore } from './IndexedDbKeyValueStore'
 export { RangeDb } from './RangeDb'
